backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
colorthief==0.2.1
webcolors==1.13
pydantic==2.5.2
aiofiles==23.2.1
python-dotenv==1.0.0
firebase-admin==6.2.0
rembg==2.0.50
